+++ stackinputhelper/db/tasks.php
@@ -1,0 +1,16 @@
+<?php
+// This file is part of Moodle - http://moodle.org/.
+
+defined('MOODLE_INTERNAL') || die();
+
+$tasks = [
+    [
+        'classname' => '\\local_stackinputhelper\\task\\cleanup_sessions',
+        'blocking' => 0,
+        'minute' => 'R',
+        'hour' => '*',
+        'day' => '*',
+        'month' => '*',
+        'dayofweek' => '*',
+    ],
+];